Bump mlx version in test and update test images (#325)

+++ tests/image_generation/helpers/image_generation_concept_test_helper.py
@@ -61,7 +61,7 @@ class ImageGenerationConceptTestHelper:
 
         finally:
             # cleanup
-            if os.path.exists(output_heatmap_path):
+            if os.path.exists(output_heatmap_path) and "MFLUX_PRESERVE_TEST_OUTPUT" not in os.environ:
                 os.remove(output_heatmap_path)
 
     @staticmethod
